Use coarse sweep as part of middle out sweep. Main is now technically any instance name. Allow 0.0 and 1.0 slices to act as io regs. Dont carry around zero clk map for delay use logic delay field.

+++ src/RAW_VHDL.py
@@ -321,7 +321,7 @@ def GET_BIN_OP_XOR_C_BUILT_IN_C_ENTITY_WIRES_DECL_AND_PROCESS_STAGES_TEXT(logic,
     # Odd, ex 5:  | | | | |
     #                 ^
     # Depends on position of middle slice
-    middle_index = latency/2
+    middle_index = int(latency/2)
     middle_slice = timing_params.slices[middle_index]
     # If slice is to left, logic is on right
     stage_for_1ll = middle_index
@@ -515,7 +515,7 @@ def GET_UNARY_OP_NOT_C_BUILT_IN_C_ENTITY_WIRES_DECL_AND_PROCESS_STAGES_TEXT(logi
     # Odd, ex 5:  | | | | |
     #                 ^
     # Depends on position of middle slice
-    middle_index = latency/2
+    middle_index = int(latency/2)
     middle_slice = timing_params.slices[middle_index]
     # If slice is to left, logic is on right
     stage_for_1ll = middle_index
@@ -588,7 +588,7 @@ def GET_BIN_OP_AND_C_BUILT_IN_C_ENTITY_WIRES_DECL_AND_PROCESS_STAGES_TEXT(logic,
     # Odd, ex 5:  | | | | |
     #                 ^
     # Depends on position of middle slice
-    middle_index = latency/2
+    middle_index = int(latency/2)
     middle_slice = timing_params.slices[middle_index]
     # If slice is to left, logic is on right
     stage_for_1ll = middle_index
@@ -663,7 +663,7 @@ def GET_BIN_OP_OR_C_BUILT_IN_C_ENTITY_WIRES_DECL_AND_PROCESS_STAGES_TEXT(logic, 
     # Odd, ex 5:  | | | | |
     #                 ^
     # Depends on position of middle slice
-    middle_index = latency/2
+    middle_index = int(latency/2)
     middle_slice = timing_params.slices[middle_index]
     # If slice is to left, logic is on right
     stage_for_1ll = middle_index
